requirements.txt
# fastapi==0.135.1
# joblib==1.5.3
# matplotlib==3.10.8
# numpy==2.4.2
# pandas==3.0.1
# pydantic==2.12.5
# pydantic_settings==2.13.1
# PyYAML==6.0.3
# scikit_learn==1.8.0

fastapi
uvicorn
joblib
numpy
pandas
scikit-learn
pydantic
pydantic-settings
PyYAML
matplotlib